readme updated, states added
